feat: create the "aggregator" module
